delta-lake/requirements.txt
# delta-spark==2.4.0
# delta-spark
delta-spark==3.0.0
deltalake
jupyterlab 
pandas
# pyspark==3.4.1
pyspark==3.5.5
apache-airflow
apache-airflow-providers-standard
apache-airflow-providers-amazon
apache-airflow-providers-apache-spark
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-sqlite
backoff
# pyspark
requests
duckdb
pypdf
pymongo